# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Froglet Node API
  version: 1.1.0
  description: >
    HTTP API for a Froglet node. Froglet is a service-execution framework; the
    node exposes two distinct ports:

    - **Provider port** (default 8080) — public requester endpoints plus
      Bearer-authenticated local provider-control endpoints. Do not expose a
      provider-control token to requesters.

    - **Runtime port** (default 8081) — Bearer-authenticated endpoints used by
      the local operator to create deals via the marketplace, manage the wallet,
      and launch workloads.

    Both ports also expose the shared `/health`, `/v1/node/capabilities`, and
    `/v1/node/identity` endpoints.
servers:
  - url: http://localhost:8080
    description: Provider port (public requester routes plus authenticated local control routes)
  - url: http://localhost:8081
    description: Runtime port (local operator, Bearer auth required)
...
security: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        Static Bearer token configured via `FROGLET_RUNTIME_AUTH_TOKEN`.
        Pass as `Authorization: Bearer <token>` on all runtime-port endpoints
        that require authentication.
    ProviderControlBearerAuth:
      type: http
      scheme: bearer
      description: >
        Static Bearer token configured via `FROGLET_PROVIDER_CONTROL_AUTH_TOKEN`.
        Pass as `Authorization: Bearer <token>` on provider-control endpoints.
...
paths:
...
  /v1/provider/services/{service_id}/files/{revision}/download:
    parameters:
      - in: path
        name: service_id
        required: true
        schema: {type: string}
      - in: path
        name: revision
        required: true
        schema: {type: string, pattern: '^[0-9a-f]{64}$'}
    get:
      summary: Download an admitted immutable file snapshot
      description: >-
        Requires the active signed file revision, enabled finite provider file limits,
        zero-fee terms, and provider access. Invite mode accepts x-froglet-access-token;
        private mode requires the provider control bearer. Reserves download count and
        full bytes before responding. Failed attempts are not refunded. Maximum 8 MiB;
        ranges are unsupported. First-party CORS only; always no-store and attachment.
      parameters:
        - in: header
          name: x-froglet-access-token
          schema: {type: string, minLength: 32, maxLength: 256}
      responses:
        '200':
          description: Exact snapshot bytes; Content-Length, ETag and X-Froglet-File-Sha256 identify them
          content:
            application/octet-stream:
              schema: {type: string, format: binary}
        '403': {description: Access or origin denied}
        '404': {description: No permitted file publication}
        '410': {description: Revision inactive, expired or paused}
        '413': {description: Relay response limit is below the file size}
        '416': {description: Range requests are unsupported}
        '429': {description: Transfer capacity or allowance exhausted}
        '503': {description: File sharing disabled or snapshot unavailable}
    head:
      summary: Inspect admitted file headers without transferring the file
      description: Performs access and lifecycle checks; does not reserve file count or bytes. Invitation admission is counted.
      responses:
        '200': {description: File headers without a body}
    options:
      summary: First-party browser download preflight
      responses:
        '204': {description: Allowed download methods and request headers}
        '403': {description: Origin is not allowed}
...